backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
beautifulsoup4==4.12.3
requests==2.31.0
langchain==0.1.0
langchain-community==0.0.13
huggingface_hub==0.20.3
pydantic==2.5.3
httpx==0.26.0
